fix: Resolves issue 1
Fixes the issue where the compiler throws a NullPointerException with
nested print expressions. This fix simply throws a semantic error to
standard error stream if this occurs. In the future, this could be
changed to simply log the data type "null" or "undefined" to standard
output instead of showing a semantic error message.